Fix Send and Sync implementations for Channel
Channel should not be usable to use with non Send/Sync messages types
because it allows transfering types across threads
This patch also relies on the auto trait implementations for Requester/Responder.

This is technically a breaking change, but I don't think it should be a major release.
This is unlikely to affect anyone, and if it does, it's probably uncovering unsoundness